+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 # Changelog
 Notable changes of the project.
 
+
+## [0.0.8] - 2024-10-17
+- Add info on Tesseract and Poppler versions on ``/version`` GET service
+- Services Docker image: ``andrealenzi/poppleract-services:0.0.8``
+- Base Docker image: ``andrealenzi/poppleract-base:24.10_5.4.1``
+  - *Python 3.12*
+  - *Leptonica 1.85.0* 
+  - *Tesseract 5.4.1*
+  - *Poppler 24.10.0*
+
+
 ## [0.0.7] - 2023-12-05
 - First release
 - Services Docker image: ``andrealenzi/poppleract-services:0.0.7``